Removed redundant variables from CaloParams.h

+++ L1Trigger/L1TCalorimeter/interface/Stage1Layer2EGammaAlgorithmImp.h
@@ -40,8 +40,8 @@ namespace l1t {
     		  const std::vector<l1t::CaloRegion> & regions)  const;
     double AssociatedJetPt(int ieta, int iphi,
 		           const std::vector<l1t::Jet> * jets) const;
-    int egSeedThreshold;
-    int jetSeedThreshold;
+    double egSeedThreshold;
+    double jetSeedThreshold;
     double egLsb, jetLsb;
     double egRelativeJetIsolationCut;
     // double HoverECut;
