+++ L1Trigger/L1TCalorimeter/interface/Stage1Layer2TauAlgorithmImp.h
@@ -49,8 +49,8 @@ namespace l1t {
     //int tauSeed;
     //double relativeIsolationCut;
     //double relativeJetIsolationCut;
-    int tauSeedThreshold;
-    int jetSeedThreshold;
+    double tauSeedThreshold;
+    double jetSeedThreshold;
     int switchOffTauIso;
     bool do2x1Algo;
     double jetLsb;
