+++ L1Trigger/L1TCalorimeter/src/JetFinderMethods.cc
@@ -26,7 +26,7 @@ namespace l1t {
     return diff;
   }
 
-  void slidingWindowJetFinder(const int jetSeedThreshold, const std::vector<l1t::CaloRegion> * regions,
+  void slidingWindowJetFinder(const double jetSeedThreshold, const std::vector<l1t::CaloRegion> * regions,
 			      std::vector<l1t::Jet> * uncalibjets)
   {
     // std::cout << "Jet Seed: " << jetSeedThreshold << std::endl;
